+++ src/OpenTelemetry.Exporter.OpenTelemetryProtocol/OtlpMetricExporterExtensions.cs
@@ -168,6 +168,15 @@ public static class OtlpMetricExporterExtensions
         Debug.Assert(metricReaderOptions != null, "metricReaderOptions was null");
         Debug.Assert(experimentalOptions != null, "experimentalOptions was null");
 
+#if NET462_OR_GREATER || NETSTANDARD2_0
+        if (exporterOptions!.Protocol == OtlpExportProtocol.Grpc &&
+            ReferenceEquals(exporterOptions.HttpClientFactory, exporterOptions.DefaultHttpClientFactory))
+        {
+            throw new NotSupportedException("OtlpExportProtocol.Grpc with the default HTTP client factory is not supported on .NET Framework or .NET Standard 2.0." +
+                "Please switch to OtlpExportProtocol.HttpProtobuf or provide a custom HttpClientFactory.");
+        }
+#endif
+
         if (!skipUseOtlpExporterRegistrationCheck)
         {
             serviceProvider!.EnsureNoUseOtlpExporterRegistrations();